Extract ProjCommitsHandler to its own package
